requirements.txt
azure-search-documents
azure-identity
httpx
fastapi
python-dotenv
gradio
openai ==1.55.1
